Correção questão 8

+++ Questão 8/Questao8/Questao8/Progressao.cs
@@ -9,7 +9,7 @@ namespace Questao8
         public int R, X, Y;
         private double Result = 0;
 
-        public int Soma()
+        public double Soma()
         {
             double D = X + ((Y - 1) * R);
             double Result = (Y * (X+D)) / 2;
